requirements.txt
Flask>=3.0
Flask-SQLAlchemy>=3.1
mysql-connector-python
Werkzeug>=3.0
sqlalchemy>=2.0
flask-jwt-extended
bcrypt
twilio
python-dotenv
flask-cors
pydantic
alembic